Fix the dompdf entry point for wp-dependencies
Last PR (#1) I updated dompdf to the latest version, but didn't update
the entry filename used here in this `wp_load_dependency` call. It was
working locally for me because I was using composer autoloading, but not
working on deployment. This fixes the path to the plugin entry so that
it will work on hosting.